[e2e] Remove unnecessary logic around copying results (#22034)
After #21905, some of the logic around handling result artifacts become unnecessary or incorrect (in generating error logs). They are removed.

+++ release/e2e.py
@@ -2207,23 +2207,22 @@ def run_test_config(
 
         out_dir = os.path.expanduser(GLOBAL_CONFIG["RELEASE_RESULTS_DIR"])
 
-        logger.info(
-            f"Moving results dir {temp_dir} to persistent location " f"{out_dir}"
-        )
+        logger.info(f"Moving results dir {temp_dir} to persistent location {out_dir}")
 
         try:
-            shutil.rmtree(out_dir)
+            # out_dir is cleared in run_e2e.sh, but it may exist when running
+            # e2e.py directly.
+            if os.path.exists(out_dir):
+                shutil.rmtree(out_dir)
+                logger.info(f"Destination {out_dir} is cleared")
         except Exception:
-            logger.exception(
-                f"Ran into error when clearing the destination dir: {out_dir}"
+            logger.info(
+                f"Ran into error when clearing the destination: {out_dir}",
+                exc_info=True,
             )
 
         try:
-            # Use distutils.dir_util.copy_tree() instead of shutil.cptree(),
-            # which allows existing output directory.
-            from distutils.dir_util import copy_tree
-
-            copy_tree(temp_dir, out_dir)
+            shutil.cptree(temp_dir, out_dir)
         except Exception:
             logger.exception(
                 "Ran into error when copying results dir to persistent "
